feat: add asset from automatic
