+++ frontend/pages/home/components/ICRC/asset/AssetDrawerMutate.tsx
@@ -2,21 +2,38 @@ import PlusIcon from "@assets/svg/files/plus-icon.svg";
 // svgs
 import {
   AssetMutationAction,
+  AssetMutationResult,
   setAccordionAssetIdx,
   setAssetMutation,
   setAssetMutationAction,
+  setAssetMutationResult,
+  setSelectedAsset,
 } from "@redux/assets/AssetReducer";
 import { useAppDispatch, useAppSelector } from "@redux/Store";
 import { useMemo } from "react";
 import { BasicDrawer } from "@components/drawer";
 import CloseAssetDrawer from "./CloseAssetDrawer";
-import AssetFormRender from "./AssetFormRender";
+import AddAssetAutomatic from "./AddAssetAutomatic";
+import AddAssetManual from "./AddAssetManual";
+import useAssetMutate from "@pages/home/hooks/useAssetMutate";
+import DialogAssetConfirmation from "./DialogAssetConfirmation";
+import { Asset } from "@redux/models/AccountModels";
+import { getAssetDetails } from "@pages/home/helpers/icrc";
+import { db } from "@/database/db";
+import { useTranslation } from "react-i18next";
 
 export default function AssetDrawerMutate() {
   const dispatch = useAppDispatch();
+  const { t } = useTranslation();
+  const { assets } = useAppSelector((state) => state.asset);
   const { accordionIndex } = useAppSelector((state) => state.asset.helper);
   const { assetMutated, assetAction } = useAppSelector((state) => state.asset.mutation);
 
+  const { newAsset, setNewAsset, setValidToken, setErrIndex, setErrToken, errIndex, errToken, validToken } =
+    useAssetMutate();
+
+  const isManual = useMemo(() => assetAction === AssetMutationAction.ADD_MANUAL, [assetAction]);
+  const isUpdate = useMemo(() => assetAction === AssetMutationAction.UPDATE, [assetAction]);
   const isDrawerOpen = useMemo(
     () => assetAction !== AssetMutationAction.NONE && assetAction !== AssetMutationAction.DELETE,
     [assetAction],
@@ -27,12 +44,41 @@ export default function AssetDrawerMutate() {
       <div className={containerStyles} onClick={onAddAsset}>
         <img src={PlusIcon} alt="plus-icon" />
       </div>
-      <BasicDrawer isDrawerOpen={isDrawerOpen}>
-        <div className={formContainerStyles}>
-          <CloseAssetDrawer onClose={onClose} isEdit={assetMutated ? true : false} />
-          <AssetFormRender />
-        </div>
-      </BasicDrawer>
+      {isDrawerOpen && (
+        <BasicDrawer isDrawerOpen={isDrawerOpen}>
+          <div className={formContainerStyles}>
+            <CloseAssetDrawer onClose={onClose} isEdit={assetMutated ? true : false} />
+
+
+            {isManual || isUpdate ? (
+              <AddAssetManual
+                errToken={errToken}
+                setErrToken={setErrToken}
+                errIndex={errIndex}
+                setErrIndex={setErrIndex}
+                validToken={validToken}
+                setValidToken={setValidToken}
+                newAsset={newAsset}
+                setNewAsset={setNewAsset}
+                asset={assetMutated}
+                addAssetToData={addAssetToData}
+              ></AddAssetManual>
+            ) : (
+              <AddAssetAutomatic
+                setNewAsset={setNewAsset}
+                newAsset={newAsset}
+                addAssetToData={addAssetToData}
+                setValidToken={setValidToken}
+                setErrToken={setErrToken}
+                errToken={errToken}
+              ></AddAssetAutomatic>
+            )}
+
+            <DialogAssetConfirmation newAsset={newAsset} />
+
+          </div>
+        </BasicDrawer>
+      )}
     </>
   );
 
@@ -47,6 +93,50 @@ export default function AssetDrawerMutate() {
 
     dispatch(setAssetMutation(undefined));
     dispatch(setAssetMutationAction(AssetMutationAction.NONE));
+  }
+
+  function isAssetAdded(address: string) {
+    return assets.find((asst: Asset) => asst.address === address) ? true : false;
+  };
+
+  async function addAssetToData() {
+    if (isAssetAdded(newAsset.address)) {
+      setErrToken(t("adding.asset.already.imported"));
+      setValidToken(false);
+    } else {
+      try {
+        dispatch(setAssetMutationResult(AssetMutationResult.ADDING));
+
+        const idxSorting = assets.length > 0 ? [...assets].sort((a, b) => b.sortIndex - a.sortIndex) : [];
+        const sortIndex = (idxSorting.length > 0 ? idxSorting[0]?.sortIndex : 0) + 1;
+
+        const updatedAsset = await getAssetDetails({
+          canisterId: newAsset.address,
+          includeDefault: true,
+          customName: newAsset.name,
+          customSymbol: newAsset.symbol,
+          supportedStandard: newAsset.supportedStandards,
+          sortIndex,
+        });
+
+        const assetToSave: Asset = { ...newAsset, ...updatedAsset, sortIndex };
+        await db().addAsset(assetToSave, { sync: true });
+
+        dispatch(setAssetMutationResult(AssetMutationResult.ADDED));
+        dispatch(setSelectedAsset(assetToSave));
+        dispatch(setAccordionAssetIdx([assetToSave.symbol]));
+
+      } catch (error) {
+        console.error("Error adding asset", error);
+        dispatch(setAssetMutationResult(AssetMutationResult.FAILED));
+      } finally {
+        setTimeout(() => {
+          dispatch(setAssetMutationResult(AssetMutationResult.NONE));
+          dispatch(setAssetMutationAction(AssetMutationAction.NONE));
+          dispatch(setAssetMutation(undefined));
+        }, 3000);
+      }
+    }
   }
 }
